Fix rake publish task

+++ Rakefile
@@ -19,15 +19,20 @@ end
 
 desc "Generate and publish blog to gh-pages"
 task :publish => [:generate] do
-  Dir.mktmpdir do |tmp|
-    system "mv _site/* #{tmp}"
-    system "git checkout -b gh-pages"
-    system "rm -rf *"
-    system "mv #{tmp}/* ."
+  if `git status -s`.empty?
+    system "git checkout -b gh-pages" unless system "git checkout gh-pages"
+    Dir.mktmpdir do |tmp|
+      system "mv _site/* #{tmp}"
+      system "rm -rf *"
+      system "mv #{tmp}/* ."
+    end
     message = "Site updated at #{Time.now.utc}"
-    system "git add ."
+    system "git add --all"
     system "git commit -am #{message.shellescape}"
-    system "git push origin gh-pages --force"
+    system "git push origin gh-pages"
     system "git checkout master"
+    system "echo publish successful"
+  else
+    system "echo publish aborted due to dirty repo"
   end
 end
